+++ examples/harmony/entry/src/main/ets/tinode/Index.ts
@@ -17,4 +17,6 @@ export * from './TinodeMessage.ts';
 export * from './TinodeStorage.ts';
 export * from './TinodeTopics.ts';
 export * from './TinodeAttachment.ts';
+export * from './TinodeStore.ts';
+export * from './TinodeStorageCache.ts';
 export * from './Tinode.ts';